core: frontend: Rename helper types file
